Refactor CLI command registry to proper architecture
ARCHITECTURAL IMPROVEMENT: Move command registry from constants to cli.py

RATIONALE:
- CLI command definitions belong in the CLI module, not constants
- constants/ should contain pure data, not utility functions
- More intuitive for developers - CLI code is in CLI module
- Cleaner separation of concerns

CHANGES:
- Moved CLI_COMMANDS dict and all utility functions to cli.py
- Removed src/mmrelay/constants/commands.py
- Updated constants/__init__.py to remove command exports
- Fixed all imports and f-string syntax errors
- Maintained all functionality while improving organization

BENEFITS:
- Intuitive code organization - CLI utilities in CLI module
- Cleaner constants module - pure constants only
- Better cohesion - command definitions with CLI logic
- Same programmatic command generation functionality
- No more hunting for hardcoded command strings

The core requirement is still fully met: programmatic command loading
that automatically displays current command syntax, but with much
better architectural organization.
