+++ src/mmrelay/matrix_utils.py
@@ -37,7 +37,6 @@ from mmrelay.config import (
     get_meshtastic_config_value,
     save_credentials,
 )
-from mmrelay.constants import msg_require_auth_login, msg_retry_auth_login
 from mmrelay.constants.app import WINDOWS_PLATFORM
 from mmrelay.constants.config import (
     CONFIG_SECTION_MATRIX,
@@ -471,7 +470,7 @@ async def connect_matrix(passed_config=None):
         # Check if matrix section exists in config
         if "matrix" not in config:
             logger.error("No Matrix authentication available. Neither credentials.json nor matrix section in config found.")
-            logger.error(f"{msg_require_auth_login()}")
+            logger.error("Please run 'mmrelay auth login' to set up credentials.json, or add matrix section to config.yaml")
             return None
 
         matrix_section = config["matrix"]
@@ -1401,7 +1400,7 @@ async def on_decryption_failure(room: MatrixRoom, event: MegolmEvent) -> None:
         f"Failed to decrypt event '{event.event_id}' in room '{room.room_id}'! "
         f"This is usually temporary and resolves on its own. "
         f"If this persists, the bot's session may be corrupt. "
-        f"{msg_retry_auth_login()}."
+Try logging in again with 'mmrelay auth login'.
     )
 
     # Attempt to request the keys for the failed event
